+++ app/src/main/java/br/com/douglas/esports/Team.kt
@@ -1,30 +1,14 @@
 package br.com.douglas.esports
 
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
 data class Team(
     val id: String,
-    val name: String
-){
-    companion object{
-        fun getMockTeam()= listOf(
-            Team(
-                id = "555",
-                name = "Fallen"
-            ),
-            Team(
-                id = "789",
-                name = "Amendoim"
-            ),
-            Team(
-                id = "123",
-                name = "Sacy"
-            ),
-            Team(
-                id = "666",
-                name = "Aspas"
-            )
-        )
+    val name: String,
+    val players: List<Player> = listOf()
+):Parcelable
 
-    }
-}
